+++ src/main/java/uk/gov/hmcts/reform/divorce/orchestration/service/CaseOrchestrationService.java
@@ -171,4 +171,6 @@ public interface CaseOrchestrationService {
 
     CcdCallbackResponse welshSetPreviousState(CcdCallbackRequest ccdCallbackRequest) throws WorkflowException;
 
+    Map<String, Object> createGeneralEmail(CcdCallbackRequest ccdCallbackRequest) throws CaseOrchestrationServiceException;
+
 }
