+++ src/main/java/uk/gov/hmcts/reform/divorce/orchestration/tasks/generalemail/GeneralEmailCoRespondentTask.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.divorce.orchestration.tasks.servicejourney.emails.general;
+package uk.gov.hmcts.reform.divorce.orchestration.tasks.generalemail;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
@@ -10,14 +10,14 @@ import uk.gov.hmcts.reform.divorce.orchestration.service.bulk.print.helper.Gener
 
 import java.util.Map;
 
-import static uk.gov.hmcts.reform.divorce.orchestration.service.bulk.print.dataextractor.CaseDataExtractor.getRespondentEmail;
+import static uk.gov.hmcts.reform.divorce.orchestration.service.bulk.print.dataextractor.CaseDataExtractor.getCoRespondentEmail;
 import static uk.gov.hmcts.reform.divorce.orchestration.service.bulk.print.helper.GeneralEmailTaskHelper.getNotRepresentedSubject;
 
 @Component
 @Slf4j
-public class GeneralEmailRespondentTask extends SendEmailTask {
+public class GeneralEmailCoRespondentTask extends SendEmailTask {
 
-    public GeneralEmailRespondentTask(EmailService emailService) {
+    public GeneralEmailCoRespondentTask(EmailService emailService) {
         super(emailService);
     }
 
@@ -28,16 +28,16 @@ public class GeneralEmailRespondentTask extends SendEmailTask {
 
     @Override
     protected Map<String, String> getPersonalisation(TaskContext context, Map<String, Object> caseData) {
-        return GeneralEmailTaskHelper.getExpectedNotificationTemplateVars(GeneralEmailTaskHelper.Party.RESPONDENT, context, caseData);
+        return GeneralEmailTaskHelper.getExpectedNotificationTemplateVars(GeneralEmailTaskHelper.Party.CO_RESPONDENT, context, caseData);
     }
 
     @Override
     protected EmailTemplateNames getTemplate() {
-        return EmailTemplateNames.GENERAL_EMAIL_RESPONDENT;
+        return EmailTemplateNames.GENERAL_EMAIL_CO_RESPONDENT;
     }
 
     @Override
     protected String getRecipientEmail(Map<String, Object> caseData) {
-        return getRespondentEmail(caseData);
+        return getCoRespondentEmail(caseData);
     }
 }
