+++ src/main/java/uk/gov/hmcts/reform/divorce/orchestration/tasks/generalemail/GeneralEmailPetitionerSolicitorTask.java
@@ -1,37 +1,37 @@
-package uk.gov.hmcts.reform.divorce.orchestration.tasks.servicejourney.emails.general;
+package uk.gov.hmcts.reform.divorce.orchestration.tasks.generalemail;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 import uk.gov.hmcts.reform.divorce.orchestration.domain.model.email.EmailTemplateNames;
 import uk.gov.hmcts.reform.divorce.orchestration.framework.workflow.task.TaskContext;
-import uk.gov.hmcts.reform.divorce.orchestration.framework.workflow.task.generics.PetitionerSendEmailTask;
+import uk.gov.hmcts.reform.divorce.orchestration.framework.workflow.task.generics.PetitionerSolicitorSendEmailTask;
 import uk.gov.hmcts.reform.divorce.orchestration.service.EmailService;
 import uk.gov.hmcts.reform.divorce.orchestration.service.bulk.print.helper.GeneralEmailTaskHelper;
 
 import java.util.Map;
 
-import static uk.gov.hmcts.reform.divorce.orchestration.service.bulk.print.helper.GeneralEmailTaskHelper.getNotRepresentedSubject;
+import static uk.gov.hmcts.reform.divorce.orchestration.service.bulk.print.helper.GeneralEmailTaskHelper.getRepresentedSubject;
 
 @Component
 @Slf4j
-public class GeneralEmailPetitionerTask extends PetitionerSendEmailTask {
+public class GeneralEmailPetitionerSolicitorTask extends PetitionerSolicitorSendEmailTask {
 
-    public GeneralEmailPetitionerTask(EmailService emailService) {
+    public GeneralEmailPetitionerSolicitorTask(EmailService emailService) {
         super(emailService);
     }
 
     @Override
     protected Map<String, String> getPersonalisation(TaskContext taskContext, Map<String, Object> caseData) {
-        return GeneralEmailTaskHelper.getExpectedNotificationTemplateVars(GeneralEmailTaskHelper.Party.PETITIONER, taskContext, caseData);
+        return GeneralEmailTaskHelper.getExpectedNotificationTemplateVars(GeneralEmailTaskHelper.Party.PETITIONER_SOLICITOR, taskContext, caseData);
     }
 
     @Override
     protected String getSubject(TaskContext context, Map<String, Object> caseData) {
-        return getNotRepresentedSubject(context);
+        return getRepresentedSubject(context, caseData);
     }
 
     @Override
     protected EmailTemplateNames getTemplate() {
-        return EmailTemplateNames.GENERAL_EMAIL_PETITIONER;
+        return EmailTemplateNames.GENERAL_EMAIL_PETITIONER_SOLICITOR;
     }
 }
